Fix to split on hiragana and katakana character changes.
Didn't do this before because the prolonged voice mark is counted
as both hiragana and katakana, throwing the entire process off.
As you don't usually want to split on that character, it's part of
the word.

I'm not sure I entirely like the solution, but it works for the
immediate future.